Fix GTE MAC register overflow causing Snake to disappear
mat_vec_mul_add_tr used 32-bit int for MAC computation:
  MAC1 = (TRX << 12) + R11*VX + R12*VY + R13*VZ

For objects with large world translations (Snake at t=[-9000,754,-9500]),
TRX<<12 is ~37M and the matrix product adds another ~36M, exceeding
INT32_MAX. PSX GTE MAC registers are 44-bit, preventing this overflow.

Fix: use long long (64-bit) for the intermediate calculation, matching
the PSX GTE's 44-bit MAC precision. Cast result to int for storage.

Also removed DG_FLAG_INVISIBLE check from renderer (PSX checks it in
DG_BoundChanl, not the renderer).

+++ port/psx/gte_math.c
@@ -79,9 +79,10 @@ static void mat_vec_mul(const MATRIX *m, long vx, long vy, long vz)
 
 static void mat_vec_mul_add_tr(const MATRIX *m, long vx, long vy, long vz)
 {
-    gte_state.MAC1 = ((int)gte_state.TRX << 12) + (int)m->m[0][0] * vx + (int)m->m[0][1] * vy + (int)m->m[0][2] * vz;
-    gte_state.MAC2 = ((int)gte_state.TRY << 12) + (int)m->m[1][0] * vx + (int)m->m[1][1] * vy + (int)m->m[1][2] * vz;
-    gte_state.MAC3 = ((int)gte_state.TRZ << 12) + (int)m->m[2][0] * vx + (int)m->m[2][1] * vy + (int)m->m[2][2] * vz;
+    /* PSX GTE MAC registers are 44-bit. Use 64-bit to prevent overflow. */
+    gte_state.MAC1 = (int)(((long long)gte_state.TRX << 12) + (long long)m->m[0][0] * vx + (long long)m->m[0][1] * vy + (long long)m->m[0][2] * vz);
+    gte_state.MAC2 = (int)(((long long)gte_state.TRY << 12) + (long long)m->m[1][0] * vx + (long long)m->m[1][1] * vy + (long long)m->m[1][2] * vz);
+    gte_state.MAC3 = (int)(((long long)gte_state.TRZ << 12) + (long long)m->m[2][0] * vx + (long long)m->m[2][1] * vy + (long long)m->m[2][2] * vz);
     gte_state.IR1 = clamp_mac_to_ir(gte_state.MAC1 >> 12);
     gte_state.IR2 = clamp_mac_to_ir(gte_state.MAC2 >> 12);
     gte_state.IR3 = clamp_mac_to_ir(gte_state.MAC3 >> 12);
